# platform/swaggo/swagger.yaml
basePath: /motogo/api/v1
definitions:
...
  handlers.BranchListItemResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      brands:
        items:
          type: string
        type: array
      displacement_ranges:
        items:
          type: string
        type: array
      establishment_type:
        type: string
      establishment_type_label:
        description: Spanish label for UI
        type: string
      franchise_id:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/handlers.LocationDTO'
      name:
        type: string
      profile_image_url:
        type: string
      status:
        type: string
    type: object
  handlers.BranchResponse:
    properties:
      _links:
        items:
          $ref: '#/definitions/handlers.Link'
        type: array
      brands:
        items:
          type: string
        type: array
      contact_phone:
        description: Representative's phone
        type: string
      displacement_ranges:
        items:
          type: string
        type: array
      establishment_type:
        type: string
      establishment_type_label:
        description: Spanish label for UI
        type: string
      franchise_id:
        type: string
      geocoding_status:
        allOf:
        - $ref: '#/definitions/handlers.GeocodingStatus'
        description: Indicates geocoding result
      id:
        type: string
      location:
        $ref: '#/definitions/handlers.LocationDTO'
      name:
        type: string
      profile_image_url:
        type: string
      status:
        type: string
    type: object
...
  handlers.CreateScheduleDetailRequest:
    properties:
      closing_time:
        description: HH:MM format, required if is_closed=false
        type: string
      day_of_week:
        description: 1=Monday to 7=Sunday
        type: integer
      is_closed:
        description: true if branch is closed on this day
        type: boolean
      opening_time:
        description: HH:MM format, required if is_closed=false
        type: string
    type: object
...
  handlers.GeocodingStatus:
    enum:
    - SUCCESS
    - FAILED
    - SKIPPED
    type: string
    x-enum-comments:
      GeocodingStatusFailed: Geocoding failed, retry via update
      GeocodingStatusSkipped: User provided coordinates, no geocoding needed
      GeocodingStatusSuccess: Coordinates were generated automatically
    x-enum-descriptions:
    - Coordinates were generated automatically
    - Geocoding failed, retry via update
    - User provided coordinates, no geocoding needed
    x-enum-varnames:
    - GeocodingStatusSuccess
    - GeocodingStatusFailed
    - GeocodingStatusSkipped
...
  handlers.Link:
    properties:
      href:
        type: string
      method:
        type: string
      rel:
        type: string
    type: object
  handlers.LocationDTO:
    properties:
      address:
        type: string
      city_id:
        type: string
      city_name:
        description: For geocoding (not persisted)
        type: string
      department_id:
        type: string
      department_name:
        description: For geocoding (not persisted)
        type: string
      latitude:
        type: number
      longitude:
        type: number
    required:
    - address
    - city_id
    - department_id
    type: object
...
  handlers.RegisterBranchRequest:
    properties:
      brands:
        items:
          type: string
        type: array
      displacement_ranges:
        items:
          type: string
        type: array
      establishment_type:
        description: WORKSHOP or STORE
        type: string
      franchise_id:
        type: string
      location:
        $ref: '#/definitions/handlers.LocationDTO'
      name:
        type: string
      profile_image_url:
        type: string
    required:
    - establishment_type
    - location
    - name
    type: object
...
  handlers.StandardResponse:
    description: Standard API response wrapper with success flag, message and optional
      data
    properties:
      code:
        description: Business message code (e.g., "MSG_BRANCH_REGISTERED")
        example: MSG_BRANCH_REGISTERED
        type: string
      data:
        description: Response data payload (type varies by endpoint)
      message:
        description: Human-readable message content
        example: Sede registrada exitosamente
        type: string
      success:
        description: Whether the operation was successful
        example: true
        type: boolean
    type: object
...
  handlers.UpdateScheduleRequest:
    properties:
      active:
        type: boolean
      end_date:
        description: YYYY-MM-DD format, null = indefinite
        type: string
      start_date:
        description: YYYY-MM-DD format
        type: string
    type: object
...
host: localhost:8085
info:
  contact:
    email: support@motogo.com
    name: Motogo API Support
    url: https://motogo.com/support
  description: API RESTful para la plataforma Motogo, implementada con arquitectura
    hexagonal y siguiendo los principios del Richardson Maturity Model (Nivel 2-3)
    con HATEOAS.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Motogo Backend API
  version: "1.0"
paths:
...
  /branches:
    get:
      description: Returns all branches owned by the authenticated representative,
        including location and brands
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.BranchListItemResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: List my branches
      tags:
      - Branch
    post:
      consumes:
      - application/json
      description: Creates a new branch (sede) for the authenticated representative
      parameters:
      - description: Branch registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterBranchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handlers.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/handlers.BranchResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Register a new branch
      tags:
      - Branch
...
  /branches/{id}/schedules:
    delete:
      description: Removes the schedule configuration for a branch
      parameters:
      - description: Branch ID (encoded)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete schedule for a branch
      tags:
      - Schedules
    get:
      description: Retrieves the schedule configuration for a branch
      parameters:
      - description: Branch ID (encoded)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get schedule for a branch
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      description: Creates a new schedule configuration for a branch
      parameters:
      - description: Branch ID (encoded)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create a schedule for a branch
      tags:
      - Schedules
    put:
      consumes:
      - application/json
      description: Updates the schedule configuration including validity dates
      parameters:
      - description: Branch ID (encoded)
        in: path
        name: id
        required: true
        type: string
      - description: Schedule update data
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Update schedule for a branch
      tags:
      - Schedules
  /branches/{id}/schedules/activate:
    put:
      description: Activates the schedule configuration for a branch
      parameters:
      - description: Branch ID (encoded)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Activate schedule for a branch
      tags:
      - Schedules
  /branches/{id}/schedules/deactivate:
    put:
      description: Deactivates the schedule configuration for a branch
      parameters:
      - description: Branch ID (encoded)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Deactivate schedule for a branch
      tags:
      - Schedules
  /branches/{id}/schedules/details:
    get:
      description: Retrieves all time slots (schedule details) for a branch
      parameters:
      - description: Branch ID (encoded)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: List time slots for a branch schedule
      tags:
      - Schedule Details
    post:
      consumes:
      - application/json
      description: Creates a new time slot (schedule detail) for a specific day of
        the week
      parameters:
      - description: Branch ID (encoded)
        in: path
        name: id
        required: true
        type: string
      - description: Schedule detail data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateScheduleDetailRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Create a time slot for a branch schedule
      tags:
      - Schedule Details
...
  /schedules/days:
    get:
      description: Returns the catalog of available days of week for schedules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get catalog of days of week
      tags:
      - Schedules
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"